Mayor cleanup of the installscript.
-Broke functionallity into functions.
-Restructured the code.

+++ install/install.sh
@@ -13,11 +13,10 @@ in_array() {
 
 add_template () {
   if [ -L ~/"$1" ] ; then
-    echo The file is already symlinked       
+    echo The file is already symlinked
   else
     if grep -q "#Template" ~/$1; then
       echo "$1" is already being sourced
-      break
     else
       if [ -f ~/"$1" ] ; then
         if [ -f templates/"$1" ]; then
@@ -30,7 +29,7 @@ add_template () {
         ln -s $dotfiles/"$1" ~/"$1"
         echo "Created symlink"
       fi
-    fi   
+    fi
   fi
 }
 
@@ -49,14 +48,14 @@ loop_dir () {
     echo "DotfilesPath: $dotfiles_path"
     for f in $(ls -a -I "." -I ".." $dotfiles_path )
     do
-      echo $f
       if ! in_array $(basename $f) ${ignorefiles[*]}; then
         if [ -f $home_path/$(basename $f) ] || [ -d $home_path/$(basename $f) ]; then
           if [ -p "$home_path/$(basename $f)"  ] || [ -L "$home_path/$(basename $f)" ]; then
-            echo "File is symlink"
+            # File is symlink
             continue
           else
-            echo "File is not symlink"
+            true
+            # File is not symlink
           fi
           echo $home_path/$(basename $f) Already exists. Overwrite y/n?
           echo  $(basename $f)
@@ -78,7 +77,6 @@ loop_dir () {
         fi
         echo "Creating symlink $home_path/$(basename $f) -> $dotfiles_path/$(basename $f)"
         ln -s $dotfiles_path/$(basename $f) $home_path/$(basename $f)
-        #ln -s $(basename $f) ~/$(basename $f)
       fi
     done
 }
@@ -94,78 +92,83 @@ install_file() {
 }
 
 clear_vim_swap() {
-  find ~ -mount -name "*~" -exec rm -rf {} \; 
+  find ~ -mount -name "*~" -exec rm -rf {} \;
+}
+
+git_setup() {
+  # Git
+  # Update repo and all submodules
+  cd ~/dotfiles
+  git submodule init
+  git submodule update
+  git pull && git submodule update --init --recursive
+  git submodule foreach --recursive git submodule update --init
+  cd ~/dotfiles/install
+}
+
+ignorefiles_setup()
+{
+  #Handle ignorefiles
+  ignorefiles=(. .. .bashrc git_configuration.sh .gitmodules install linstall README.md tmp.tmp deploy .git .ssh .config .local *.~* )
+
+  if in_array xfce ${desktop_managers[*]}; then
+    ignorefiles+=(".mateconf")
+  fi
+
+  if [[ "$linuxenv" == cygwin ]]
+  then
+    ignorefiles+=(.xchat2 .mateconf .config .local Desktop .komodoedit)
+    rm -rf ~/.bash_profile
+  fi
+
+  # cli mode preset (preset for commandline)
+  if [ "$1" == climode ]; then
+    ignorefiles+=(.xchat2 .mateconf .config .local Desktop .komodoedit share applications icons)
+  fi
+
+  echo "The contents of the ignorefiles array: ${ignorefiles[*]}"
 }
 
 
+# Source my own .bashrc after the systemone if it exists otherwise symlink the dotfiles one to the homedir.
+
+add_template ".bashrc"
+add_template ".profile"
 
 
 
 #Preparations
 clear_vim_swap
+git_setup
+desktop_managers=($(find /usr/share/xsessions -name "*.desktop" -exec basename "{}" .desktop ";"))
+ignorefiles_setup
+
+if [[ $(uname -a) == *CYGWIN* ]]; then
+  linuxenv=cygwin
+fi
 
 
-
-# Git
-
-# Update repo and all submodules
-cd ~/dotfiles
-git submodule init
-git submodule update
-git pull && git submodule update --init --recursive
-git submodule foreach --recursive git submodule update --init
-cd ~/dotfiles/install
+#If update only the git setup will be done
 
 if [ "$1" == update ]; then
   exit
 fi
 
+#Installing and copying files
+
 # Installing dotfiles.sh file
 install_file templates/dotfiles.sh /etc/profile.d/
 
-
-desktop_managers=($(find /usr/share/xsessions -name "*.desktop" -exec basename "{}" .desktop ";"))
-echo ${desktop_managers[*]}
-
-ignorefiles=(. .. .bashrc git_configuration.sh .gitmodules install linstall README.md tmp.tmp deploy .git .ssh .config .local *.~* )
-
-if in_array xfce ${desktop_managers[*]}; then
-  ignorefiles+=(".mateconf")
-fi
-
-if in_array mate ${desktop_managers[*]}; then
-  #ignorefiles+=(".")
-  echo "Inarray mate"
-fi
-
-if [[ $(uname -a) == *CYGWIN* ]]; then
-  linuxenv=cygwin
-  echo $linuxenv
-fi
-
-if [[ "$linuxenv" == cygwin ]]
-then
-  ignorefiles+=(.xchat2 .mateconf .config .local Desktop .komodoedit)
-  rm -rf ~/.bash_profile
-fi
-
-# cli mode preset (preset for commandline)
-if [ "$1" == climode ]; then
-  ignorefiles+=(.xchat2 .mateconf .config .local Desktop .komodoedit)
-fi
-
-echo "The contents of the ignorefiles array: ${ignorefiles[*]}"
-
-# Source my own .bashrc after the systemone if it exists otherwise symlink the dotfiles one to the homedir.
-add_template ".bashrc"
-add_template ".profile"
-
+# Setup symlinks between dotfiles and home directory
 echo "Going through all files in the dotfiles dir."
 
 loop_dir
 loop_dir .config
-loop_dir .local/share/icons
-loop_dir .local/share/applications
+
+if [[ "$linuxenv" != cygwin ]]; then
+  loop_dir .local/share/icons
+  loop_dir .local/share/applications
+fi
 
 #Handle special files
 #ssh.config
@@ -174,12 +177,12 @@ ln -s ~/dotfiles/.ssh/config ~/.ssh/config
 #Dependencies
 
 #Vim TabBar
-if commandExists sudo; then
-  sudo apt-get install ctags
-else
-  apt-get install ctags
-fi
 
-#Download .simplenoterc file to homefolder.
-#scp root@carlb.dyndns.org:~/.simplenoterc ~/
-#chmod 600 ~/.simplenoterc
+
+if commandExists apt-get; then
+  if commandExists sudo; then
+    sudo apt-get install ctags
+  else
+    apt-get install ctags
+  fi
+fi
